% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{add_junction_heights}
\alias{add_junction_heights}
\title{add heights from dtm}
\usage{
add_junction_heights(junctions, dtm)
}
\description{
add heights from dtm
}
\keyword{internal}
