% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{add_outfall_connection}
\alias{add_outfall_connection}
\title{add conduits to connect network to outfalls}
\usage{
add_outfall_connection(
  outfalls,
  junctions,
  conduits_sf,
  min_slope,
  junc_depth,
  lim
)
}
\description{
add conduits to connect network to outfalls
}
\keyword{internal}
