% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{add_point_to_street}
\alias{add_point_to_street}
\title{add a new point to an existing street}
\usage{
add_point_to_street(point_to_add, street_vertices)
}
\description{
add a new point to an existing street
}
\keyword{internal}
