% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{adjustSlopesAndJuncDepths}
\alias{adjustSlopesAndJuncDepths}
\title{slope and junction depth adjustment algorithm}
\usage{
adjustSlopesAndJuncDepths(
  conduits,
  junctions,
  outfall,
  min_slope,
  max_slope,
  min_junc_depth,
  mean_junc_depth,
  max_junc_depth
)
}
\description{
slope and junction depth adjustment algorithm
}
\keyword{internal}
