% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{backtracking_paths}
\alias{backtracking_paths}
\title{track path from junction (nearest to outfall) to start}
\usage{
backtracking_paths(
  start,
  outfalls,
  junctions,
  conduits_sf,
  next_to_seen = list()
)
}
\description{
track path from junction (nearest to outfall) to start
}
\keyword{internal}
