% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{breaks_in_network}
\alias{breaks_in_network}
\title{breaks in network}
\usage{
breaks_in_network(
  break_closed_loops,
  break_loops,
  breaks_at_hills,
  delete_disconnected,
  junctions,
  outfalls,
  conduits_sf
)
}
\description{
breaks in network
}
\keyword{internal}
