% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{calculate_new_bottom_height}
\alias{calculate_new_bottom_height}
\title{calculate new bottom height}
\usage{
calculate_new_bottom_height(length, slope, FromHeight)
}
\description{
calculate new bottom height
}
\keyword{internal}
