% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{correct_artificial_outfalls}
\alias{correct_artificial_outfalls}
\title{correct artificial outfalls}
\usage{
correct_artificial_outfalls(junctions, conduits_sf)
}
\description{
correct artificial outfalls
}
\keyword{internal}
