% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{correct_sinks}
\alias{correct_sinks}
\title{correct sinks}
\usage{
correct_sinks(
  junctions,
  conduits_sf,
  outfalls,
  short_cut_sinks,
  direct_drainage_sinks,
  ds
)
}
\description{
correct sinks
}
\keyword{internal}
