% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{create_conduits_along_topography}
\alias{create_conduits_along_topography}
\title{create conduits along the heights of the junctions}
\usage{
create_conduits_along_topography(vertices, main)
}
\description{
create conduits along the heights of the junctions
}
\keyword{internal}
