% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_drainage_network.R
\name{create_drainage_network}
\alias{create_drainage_network}
\title{Create a drainage network including junctions, conduits and outfalls.}
\usage{
create_drainage_network(
  streets,
  dtm,
  outfalls,
  crs_default,
  buffer,
  snap_dist,
  epsilon,
  lim,
  min_junc_depth,
  mean_junc_depth,
  max_junc_depth,
  min_slope,
  max_slope,
  ds,
  stepwise,
  break_closed_loops,
  delete_disconnected,
  breaks_at_hills,
  break_loops,
  short_cut_sinks,
  direct_drainage_sinks
)
}
\arguments{
\item{streets}{Street polyline data of class sf. On basis of the street 
polylines the drainage network is setup. Street polyline data can be extracted 
for example from openstreetmaps.}

\item{dtm}{Digital terrain information either of class RasterLayer or as a 
triangulated grid of class sf. Surface height data is required to add heights 
to junctions and to define the flow directions in the conduits subsequently.}

\item{outfalls}{Outfall point data of class sf. The network drains towards 
these points.}

\item{crs_default}{A proj4string (if dtm is of class RasterLayer) or EPSG Code
to specify the default coordinate reference system. The measure of the default 
crs must be in m.}

\item{buffer}{Minimum distance between street nodes (unit: meter). If distance 
is smaller street nodes are summarized. A parameter for quality checks on 
street polyline data.}

\item{snap_dist}{Minimum distance between a polyline and an ending node (unit: 
meter). If distance is smaller street polylines are connected. A parameter 
for quality checks on street polyline data.}

\item{epsilon}{Distance to ideal line between start and end point (unit: meter). 
Parameter from the Douglas Peucker Algorithm that is needed to simplify the 
shape of polylines. See \link[kmlShape]{DouglasPeuckerEpsilon} for further 
information.}

\item{lim}{The maximum distance between two junctions (unit: meter). Can be 
interpreted as maximum pipe length. To avoid very short pipes this values can 
be exceeded occasionally.}

\item{min_junc_depth}{Minimum junction depth (unit: meter).}

\item{mean_junc_depth}{Mean junction depth (unit: meter).}

\item{max_junc_depth}{Maximum junction depth (unit: meter).}

\item{min_slope}{Minimum slope value (unit: -)}

\item{max_slope}{Maximum slope value (unit: -)}

\item{ds}{Threshold for local sinks in the network (unit: meter). If the depth 
of a local sink is greater than ds (depression storage). A shortcut is 
implemented either to the nearest outfall if direct_drainage_sinks is set 
TRUE or to the nearest lower junction if short_cut_sinks is set TRUE.}

\item{stepwise}{Stepwise construction of the network if set to TRUE. When a 
stepwise construction is chosen, at first a base network including junctions 
at start, end and crossing fix points is computed. Then the local sink nodes 
are corrected.}

\item{break_closed_loops}{-}

\item{delete_disconnected}{-}

\item{breaks_at_hills}{If set to TRUE: The drainage network is forked at 
junctions that have incoming and more than one outgoing pipe. Parameter to 
increase the linearity of the network.}

\item{break_loops}{If set to TRUE: The drainage network is forked at junctions 
that have no incoming but more than one outgoing pipe. Parameter to increase 
the linearity of the network.}

\item{short_cut_sinks}{If set to TRUE: drainage of local sinks that are deeper 
than ds towards the nearest lower junction.}

\item{direct_drainage_sinks}{If set to TRUE (default): drainage of local sinks 
that are deeper than ds towards the nearest outfall node.}
}
\value{
A list of sf objects junctions, conduits and outfalls.
}
\description{
Fully automated generation of a drainage network inclduding junctions, conduits 
and outfalls based on open source spatial data. Requires street polylines, a 
digital terrain model, outfall point data and some user defined parameters as 
input. Returns a list of the sf objects junctions, conduits and outalls that 
can be converted to a SWMM *.inp file using swmmr functions afterwards.
}
