% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{create_junctions}
\alias{create_junctions}
\title{create junctions based on street polylines}
\usage{
create_junctions(
  streets,
  buffer,
  snap_dist,
  epsilon,
  lim,
  junc_depth,
  dtm,
  crs_default,
  pre_def_junctions,
  pre_def_conduits
)
}
\description{
create junctions based on street polylines
}
\keyword{internal}
