% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{create_linestrings_from_single_layer}
\alias{create_linestrings_from_single_layer}
\title{create polylines from point data}
\usage{
create_linestrings_from_single_layer(new_segment)
}
\description{
create polylines from point data
}
\keyword{internal}
