% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{delete_disconnected_network_parts}
\alias{delete_disconnected_network_parts}
\title{delete junctions and conduits that are not connected to an outfall}
\usage{
delete_disconnected_network_parts(junctions, conduits, outfalls)
}
\description{
delete junctions and conduits that are not connected to an outfall
}
\keyword{internal}
