% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{linear_path}
\alias{linear_path}
\title{track path from outfall to crossing}
\usage{
linear_path(next_to, outfalls, junctions, conduits_sf)
}
\description{
track path from outfall to crossing
}
\keyword{internal}
