% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{new_point}
\alias{new_point}
\title{Finds point in distance di from point p0 in direction of point p1}
\usage{
new_point(p0, p1, di)
}
\description{
Finds point in distance di from point p0 in direction of point p1
}
\keyword{internal}
