% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{pipeSizingAlgorithm}
\alias{pipeSizingAlgorithm}
\title{Pipe diameter sizing algorithm}
\usage{
pipeSizingAlgorithm(
  inp,
  path_out,
  conduits_sf,
  junctions,
  outfalls,
  target,
  DN_start = 300,
  rpt_filename = "artificial_SWMM_model.rpt",
  inp_filename = "artificial_SWMM_model.inp",
  link_filename = "links_artificial_SWMM_format.shp"
)
}
\arguments{
\item{inp}{A object of class inp. Containing a complete SWMM model.}

\item{path_out}{A directory path to write the results to.}

\item{conduits_sf}{A sf object containing polyline information.}

\item{junctions}{A sf object containing junction point information.}

\item{outfalls}{A sf object containing outfall point informations.}

\item{target}{Either node_flooding or conduit_surcharge.}

\item{DN_start}{Initial pipe diameter (unit: mm).}

\item{rpt_filename}{File name for SWMM's report file.}

\item{inp_filename}{File name for SWMM's input file.}

\item{link_filename}{File name for a shp file containing the sized conduits.}
}
\value{
A sf object containing polylines with all attributes needed for the conduit section in SWMM's input file.
}
\description{
The conduit diameters are calculated and designed for a user defined storm event. Initially the sizing algorithm runs a SWMM model with all conduits having a diameter of 0.3 m by default (or smaller if defined). After each iteration, it is checked whether conduits still surcharge. If so, diameters of the surcharging conduits are increased by a nominal diameter. The algorithm stops if no conduit surcharge occurs any more. The final inp file is stored.
}
