% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drainage_network_workflow.R
\name{plot_cross_section_junctions}
\alias{plot_cross_section_junctions}
\title{cross section plot when adjusting pipe slopes and junction depths}
\usage{
plot_cross_section_junctions(junctions, conduits, connections)
}
\description{
cross section plot when adjusting pipe slopes and junction depths
}
\keyword{internal}
