% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drainage_network_workflow.R
\name{plot_disconnected_conduits}
\alias{plot_disconnected_conduits}
\title{plot conduits that are not connected to an outfall}
\usage{
plot_disconnected_conduits(
  junctions,
  conduits,
  outfalls,
  col_arrow,
  main = "conduits disconnected from outfalls"
)
}
\description{
plot conduits that are not connected to an outfall
}
\keyword{internal}
