% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drainage_network_workflow.R
\name{plot_drainage_network}
\alias{plot_drainage_network}
\title{plot of drainage network during creation process}
\usage{
plot_drainage_network(coords, outfalls, conduits_sf, main, col_arrow)
}
\description{
plot of drainage network during creation process
}
\keyword{internal}
