% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drainage_network_workflow.R
\name{plot_drainage_network_pipe_diameters}
\alias{plot_drainage_network_pipe_diameters}
\title{plot pipe diameters}
\usage{
plot_drainage_network_pipe_diameters(
  junctions,
  outfalls,
  conduits_sf,
  main,
  DN_m
)
}
\description{
plot pipe diameters
}
\keyword{internal}
