% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drainage_network_workflow.R
\name{plot_path_to_outfall}
\alias{plot_path_to_outfall}
\title{plot catchment of outfalls}
\usage{
plot_path_to_outfall(junctions_u, outfalls, oid, conduits_u, main, col_arrow)
}
\description{
plot catchment of outfalls
}
\keyword{internal}
