% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drainage_network_workflow.R
\name{plot_short_cuts}
\alias{plot_short_cuts}
\title{plot short cuts}
\usage{
plot_short_cuts(junctions, conduits, outfalls, col_arrow)
}
\description{
plot short cuts
}
\keyword{internal}
