% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drainage_network_workflow.R
\name{plot_slopes}
\alias{plot_slopes}
\title{plot slopes}
\usage{
plot_slopes(
  junctions,
  outfalls,
  conduits_sf,
  min_slope,
  max_slope,
  main,
  roundValue = 3
)
}
\description{
plot slopes
}
\keyword{internal}
