% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drainage_network_workflow.R
\name{plot_tagged_junctions}
\alias{plot_tagged_junctions}
\title{plot tagged junctions}
\usage{
plot_tagged_junctions(junctions, outfalls, conduits_sf, main)
}
\description{
plot tagged junctions
}
\keyword{internal}
