% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{tag_nodes}
\alias{tag_nodes}
\title{tag nodes during creation of junctions}
\usage{
tag_nodes(street_vertices)
}
\description{
tag nodes during creation of junctions
}
\keyword{internal}
