% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{thiessen_to_subcatchments}
\alias{thiessen_to_subcatchments}
\title{combine thiessen polygons and land use information}
\usage{
thiessen_to_subcatchments(
  landuse_sf = NULL,
  landuse_classes = NULL,
  re = NULL,
  area_re_pixel = NULL,
  re_classes = NULL,
  thiessen_polygons,
  width = NULL,
  slope = NULL,
  soil = NULL
)
}
\arguments{
\item{landuse_sf}{A sf object containing polygons with land use informations as attributes.}

\item{landuse_classes}{A data frame specifying the percentage of imperviousness for landuse_sf.}

\item{re}{Alternative to landuse_sf: A raster file containing landuse information. For example derived from RapidEye Satelite Imagery.}

\item{area_re_pixel}{A data frame adding attributes to the re raster pixel values.}

\item{re_classes}{A data frame specifying the percentage of imperviousness for area_re_pixel.}

\item{thiessen_polygons}{A sf object containing Thiessen polygons with junctions as centroids. Run function \link[urbandrain]{thiessenpolygons}.}

\item{width}{not implemented yet}

\item{slope}{not implemented yet}

\item{soil}{not implemented yet}
}
\value{
A sf object including all information needed for the subcatchment input section in SWMM's input file.
}
\description{
combine thiessen polygons and land use information
}
\keyword{internal}
