% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{thiessenpolygons}
\alias{thiessenpolygons}
\title{Create Thiessen polygons}
\usage{
thiessenpolygons(points, boundary_polygon)
}
\arguments{
\item{points}{Centroid points for Thiessen tesselation. Here junction nodes.}

\item{boundary_polygon}{model boundary polygon to cut Thiessen polygons.}
}
\value{
A sf object containing Thiessen polygons.
}
\description{
Simplified approach to define recharging surfaces for SWMM model.
}
