% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage_network_helper_functions.R
\name{track_outfall_to_start_connectivity}
\alias{track_outfall_to_start_connectivity}
\title{track paths from all outfalls to start}
\usage{
track_outfall_to_start_connectivity(outfalls, junctions, conduits_sf)
}
\description{
track paths from all outfalls to start
}
\keyword{internal}
